+++ examples/serial_printline.rs
@@ -2,7 +2,7 @@
 
 #[cfg(unix)]
 use mio::{Events, Interest, Poll, Token};
-use serial_io::{build, Serial, TTYPort};
+use serial_io::{build, Serial};
 use std::env;
 use std::io;
 use std::io::Read;
@@ -30,8 +30,7 @@ pub fn main() {
     // Create the listener
     println!("Opening {} at 9600,8N1", &tty_path);
     let builder = build(tty_path, 9600);
-    let port = TTYPort::open(&builder).unwrap();
-    let mut rx = Serial::from_serial(port).unwrap();
+    let mut rx = Serial::from_builder(&builder).unwrap();
 
     poll.registry()
         .register(&mut rx, SERIAL_TOKEN, Interest::READABLE)